+++ app/components/dato/code.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 class Dato::Code < Dato::DastNode
-  def initialize(node, root)
-    super(node, "code", root)
+  def initialize(node, root, parent = nil)
+    super(node, "code", root, parent)
   end
 
   def generated_tag