+++ app/components/dato/dast_node.rb
@@ -2,8 +2,8 @@
 
 module Dato
   class DastNode < Node
-    def initialize(node, type, root = nil)
-      super(node, root)
+    def initialize(node, type, root = nil, parent = nil)
+      super(node, root, parent)
       unless node.type == type
         raise ArgumentError.new("The node type is '#{node.type}' instead of '#{type}'")
       end