+++ app/components/dato/heading.rb
@@ -1,11 +1,15 @@
 # frozen_string_literal: true
 
 class Dato::Heading < Dato::DastNode
-  def initialize(node, root)
-    super(node, "heading", root)
+  def initialize(node, root, parent = nil)
+    super(node, "heading", root, parent)
+  end
+
+  def level
+    @node.level
   end
 
   def generated_tag
-    "h#{@node.level}"
+    "h#{level}"
   end
 end